:recycle: Extract get_tx_query method into DefaultTransactionObfuscation
