+++ client-core/src/cipher/default.rs
@@ -2,6 +2,8 @@ use super::sgx::EnclaveAttr;
 use crate::TransactionObfuscation;
 use chain_core::tx::data::TxId;
 use chain_core::tx::{TxAux, TxWithOutputs};
+use client_common::tendermint::types::AbciQueryExt;
+use client_common::tendermint::Client;
 use client_common::SECP;
 use client_common::{
     Error, ErrorKind, PrivateKey, Result, ResultExt, SignedTransaction, Transaction,
@@ -53,6 +55,31 @@ impl DefaultTransactionObfuscation {
         DefaultTransactionObfuscation {
             tqe_address,
             tqe_hostname: dns_name,
+        }
+    }
+
+    /// Get DefaultTransactionObfuscation from txquery call to Tendermint client
+    pub fn from_tx_query<C>(tendermint_client: &C) -> Result<DefaultTransactionObfuscation>
+    where
+        C: Client,
+    {
+        let result = tendermint_client.query("txquery", &[])?.bytes()?;
+        let address = std::str::from_utf8(&result).chain(|| {
+            (
+                ErrorKind::ConnectionError,
+                "Unable to decode txquery address",
+            )
+        })?;
+        if let Some(hostname) = address.split(':').next() {
+            Ok(DefaultTransactionObfuscation::new(
+                address.to_string(),
+                hostname.to_string(),
+            ))
+        } else {
+            Err(Error::new(
+                ErrorKind::ConnectionError,
+                "Unable to decode txquery address",
+            ))
         }
     }
 }
